re organize peak localization
