+++ src/spikeinterface/sortingcomponents/peak_localization/method_list.py
@@ -1,0 +1,10 @@
+from .center_of_mass import LocalizeCenterOfMass
+from .monopolar import LocalizeMonopolarTriangulation
+from .grid import LocalizeGridConvolution
+
+_methods_list = [
+    LocalizeCenterOfMass,
+    LocalizeMonopolarTriangulation,
+    LocalizeGridConvolution
+]
+localization_methods = {m.name: m for m in _methods_list}
